+++ frontend/package.json
@@ -40,6 +40,6 @@
   },
   "devDependencies": {
     "node-sass": "^7.0.1",
-    "sass": "^1.50.1"
+    "sass": "^1.51.0"
   }
 }
